+++ tiny/src/components/image-gallery/manifest-actions/manifest-editor/editor-primitives/language-map-editor.tsx
@@ -3,7 +3,7 @@ import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
-import type { LanguageMap } from '../manifest-editor-types';
+import type { LanguageMap } from '../manifest-rights-options';
 import { 
   Select, 
   SelectContent,
